+++ src/cache/category/useGetCategories.ts
@@ -1,5 +1,5 @@
 import { useQuery } from '@tanstack/react-query';
-import { getCategories } from 'src/api/categories/getCategories';
+import { getCategories } from 'src/api/category/getCategories';
 
 import { queryKeys } from '../keys';
