+++ src/cache/category/useGetCategoryWords.ts
@@ -1,13 +1,12 @@
 import { useQuery } from '@tanstack/react-query';
-import { getCategoryWords } from 'src/api/words/getCategoryWords';
+import { getCategoryWords } from 'src/api/category/getWords';
 
 import { queryKeys } from '../keys';
 
-export const useGetCategoryWords = (categoryId: number) => {
+export const useGetCategoryWords = (categoryId?: number) => {
   return useQuery({
     queryKey: queryKeys.category.words(categoryId).queryKey,
-    queryFn: ({ queryKey: [, , id] }) => getCategoryWords(id),
+    queryFn: ({ queryKey: [, , { id }] }) => getCategoryWords(id),
     staleTime: Infinity,
-    enabled: !isNaN(categoryId),
   });
 };
