+++ src/cache/game/useGetGameWord.ts
@@ -1,11 +1,11 @@
 import { useQuery } from '@tanstack/react-query';
-import { getGameWord } from 'src/api/words/getGameWord';
+import { getGameWord } from 'src/api/game/getGameWord';
 
 import { queryKeys } from '../keys';
 
 export const useGetGameWord = () => {
   return useQuery({
-    queryKey: queryKeys.word.game.queryKey,
+    queryKey: queryKeys.game.word.queryKey,
     queryFn: () => getGameWord(),
   });
 };
